Move Dictionary to top level module
